Cleanup Ice/helloUI
