+++ swift/Ice/helloUI/Client.swift
@@ -7,11 +7,10 @@ import Ice
 import PromiseKit
 
 class Client: ObservableObject {
-    var helloPrx: HelloPrx!
-    var communicator: Ice.Communicator!
+    private var helloPrx: HelloPrx!
+    private var communicator: Ice.Communicator
 
-    let hostnameKey = "hostnameKey"
-
+    // For Button view bindings
     @Published var helloEnabled: Bool = true
     @Published var shutdownEnabled = true
     @Published var flushEnabled = false
@@ -23,16 +22,16 @@ class Client: ObservableObject {
         }
     }
 
-    // Error bindings
+    // For Error bindings
     @Published var error: Error?
     @Published var showingError = false
 
-    // Status bindings
+    // For Status bindings
     @Published var isSpinning = false
     @Published var statusMessage = ""
 
-    func configure(comm: Ice.Communicator) {
-        communicator = comm
+    init() {
+        communicator = Client.loadCommunicator()
     }
 
     func exception(_ err: Error) {
@@ -40,16 +39,69 @@ class Client: ObservableObject {
         showingError = true
     }
 
+    func flushBatch() {
+        firstly {
+            helloPrx.ice_flushBatchRequestsAsync()
+        }.done {
+            self.flushBatchSend()
+        }.catch { error in
+            self.exception(error)
+        }
+    }
+
+    private class func loadCommunicator() -> Ice.Communicator {
+        var initData = Ice.InitializationData()
+        let properties = Ice.createProperties()
+
+        properties.setProperty(key: "Ice.Plugin.IceDiscovery", value: "0")
+        properties.setProperty(key: "Ice.Plugin.IceSSL", value: "1")
+        properties.setProperty(key: "IceSSL.CheckCertName", value: "0")
+        properties.setProperty(key: "IceSSL.DefaultDir", value: "certs")
+        properties.setProperty(key: "IceSSL.CAs", value: "cacert.der")
+        properties.setProperty(key: "IceSSL.CertFile", value: "client.p12")
+        properties.setProperty(key: "IceSSL.Password", value: "password")
+        initData.properties = properties
+        do {
+            return try Ice.initialize(initData)
+        } catch {
+            print(error)
+            fatalError()
+        }
+    }
+
+    func flushBatchSend() {
+        flushEnabled = false
+        statusMessage = "Flushed batch requests"
+    }
+
+    func queuedRequest(_ name: String) {
+        flushEnabled = true
+        statusMessage = "Queued \(name) request"
+    }
+
+    func ready() {
+        helloEnabled = true
+        shutdownEnabled = true
+        statusMessage = "Ready"
+        isSpinning = false
+    }
+
+    func requestSent() {
+        if proxySettings.deliveryMode == .Twoway || proxySettings.deliveryMode == .TwowaySecure {
+            statusMessage = "Waiting for response"
+        } else {
+            ready()
+        }
+    }
+
     func sayHello() {
         do {
-            if helloPrx == nil {
-                return
-            }
             let delay = Int32(proxySettings.delay)
             if proxySettings.deliveryMode != .OnewayBatch,
                proxySettings.deliveryMode != .OnewaySecureBatch,
                proxySettings.deliveryMode != .DatagramBatch
             {
+                // Non Batched requests get sent asynchronously
                 var response = false
                 firstly {
                     helloPrx.sayHelloAsync(delay, sentOn: DispatchQueue.main) { _ in
@@ -65,6 +117,7 @@ class Client: ObservableObject {
                     self.exception(error)
                 }
             } else {
+                // Batch requests get enqueued locally thus this call is non blocking
                 try helloPrx.sayHello(delay)
                 queuedRequest("hello")
             }
@@ -73,25 +126,15 @@ class Client: ObservableObject {
         }
     }
 
-    func flushBatch() {
-        if helloPrx == nil {
-            return
-        }
-        firstly {
-            helloPrx.ice_flushBatchRequestsAsync()
-        }.done {
-            self.flushBatchSend()
-        }.catch { error in
-            self.exception(error)
-        }
+    func sendingRequest() {
+        helloEnabled = false
+        shutdownEnabled = false
+        statusMessage = "Sending request"
+        isSpinning = true
     }
 
     func shutdown() {
         do {
-            if helloPrx == nil {
-                return
-            }
-
             if proxySettings.deliveryMode != .OnewayBatch,
                proxySettings.deliveryMode != .OnewaySecureBatch,
                proxySettings.deliveryMode != .DatagramBatch
@@ -123,10 +166,7 @@ class Client: ObservableObject {
     }
 
     func updateProxy() throws {
-        let mode = proxySettings.deliveryMode
-
         if proxySettings.hostname == "" {
-            helloPrx = nil
             return
         }
 
@@ -134,10 +174,9 @@ class Client: ObservableObject {
             "ssl -h \"\(proxySettings.hostname)\" -p 10001:" +
             "udp -h \"\(proxySettings.hostname)\" -p 10000"
 
-        UserDefaults.standard.set(proxySettings.hostname, forKey: hostnameKey)
         helloPrx = uncheckedCast(prx: try communicator.stringToProxy(s)!, type: HelloPrx.self)
 
-        switch mode {
+        switch proxySettings.deliveryMode {
         case .Twoway:
             helloPrx = helloPrx.ice_twoway()
         case .TwowaySecure:
@@ -159,50 +198,5 @@ class Client: ObservableObject {
         if proxySettings.timeout != 0 {
             helloPrx = helloPrx.ice_invocationTimeout(Int32(proxySettings.timeout))
         }
-    }
-
-    func sendingRequest() {
-        helloEnabled = false
-        shutdownEnabled = false
-        statusMessage = "Sending request"
-        isSpinning = true
-    }
-
-    func requestSent() {
-            if let connection = helloPrx.ice_getCachedConnection() {
-                do {
-                    // Loop through the connection informations until we find an IPConnectionInfo class.
-                    for info in sequence(first: try connection.getInfo(), next: { $0.underlying }) {
-                        if let ipinfo = info as? IPConnectionInfo {
-                            proxySettings.hostname = ipinfo.remoteAddress
-                            break
-                        }
-                    }
-                } catch {
-                    // Ignore.
-                }
-            }
-        if proxySettings.deliveryMode == .Twoway || proxySettings.deliveryMode == .TwowaySecure {
-            statusMessage = "Waiting for response"
-        } else {
-            ready()
-        }
-    }
-
-    func ready() {
-        helloEnabled = true
-        shutdownEnabled = true
-        statusMessage = "Ready"
-        isSpinning = false
-    }
-
-    func queuedRequest(_ name: String) {
-        flushEnabled = true
-        statusMessage = "Queued \(name) request"
-    }
-
-    func flushBatchSend() {
-        flushEnabled = false
-        statusMessage = "Flushed batch requests"
     }
 }
